requirements.txt
streamlit
google-genai
python-dotenv
beautifulsoup4
requests
pandas
playwright